New interface

+++ src/main/java/Quiniela/service/MatchServiceImpl.java
@@ -1,6 +1,7 @@
 package quiniela.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.stereotype.Service;
@@ -15,6 +16,8 @@ import quiniela.utils.CVSParser;
 
 import javax.annotation.PostConstruct;
 import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -41,22 +44,29 @@ public class MatchServiceImpl implements MatchService {
 
     final static private Integer CSV_MAX_VALUES_MATCHES = 6;
 
+    @Value("${clean_and_build}")
+    Boolean cleanAndBuild;
 
     @PostConstruct
     private void init() {
-        matchRepository.deleteAll();
-        List<String> values = CVSParser.ParseMatches("matches.cvs");
-        List<Match> inserMatches = new ArrayList<>();
-        for (int i = CSV_MAX_VALUES_MATCHES; i < values.size(); i += CSV_MAX_VALUES_MATCHES) {
-            Match match = new Match();
-            match.setId(Long.parseLong(values.get(i)));
-            match.setDate(LocalDateTime.parse(values.get(i + 1), formatter));
-            match.setHomeTeam(values.get(i + 2));
-            match.setVisitorTeam(values.get(i + 3));
-            match.setTypeMatch(TypeMatch.valueOf(values.get(i + 5)));
-            inserMatches.add(match);
+        if(cleanAndBuild) {
+            matchRepository.deleteAll();
+            List<String> values = CVSParser.ParseMatches("matches.cvs");
+            List<Match> inserMatches = new ArrayList<>();
+            for (int i = CSV_MAX_VALUES_MATCHES; i < values.size(); i += CSV_MAX_VALUES_MATCHES) {
+                Match match = new Match();
+
+                match.setId(Long.parseLong(values.get(i)));
+                ZonedDateTime zoneTime = ZonedDateTime.of(LocalDateTime.parse(values.get(i + 1), formatter), ZoneId.of("UTC+02:00"));
+                match.setDate(zoneTime.toInstant().toEpochMilli());
+
+                match.setHomeTeam(values.get(i + 2));
+                match.setVisitorTeam(values.get(i + 3));
+                match.setTypeMatch(TypeMatch.valueOf(values.get(i + 5)));
+                inserMatches.add(match);
+            }
+            matchRepository.saveAll(inserMatches);
         }
-        matchRepository.saveAll(inserMatches);
     }
